requirements.txt
altgraph==0.17.3
pycryptodome==3.17
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
